requirements/libraries.txt
# These packages are installed from PyPI.
backports.csv==1.0.7
base36==0.1.1
beautifulsoup4==4.8.2
boto3==1.7.80
dj-database-url==0.5.0
djangorestframework==3.11.2
django-cors-headers==3.9.0
django-csp==3.7
django-elasticsearch-dsl==7.1.4
django-extensions==2.1.3
django-flags==5.0.8
django-formtools==2.3
django-health-check==3.16.5
django-localflavor==2.2
django-mptt==0.12.0
django-storages==1.11.1
django-treebeard==4.2.0
django-url-or-relative-url-field==0.1.2
django-watchman==0.15.0
djangosaml2==1.3.5
edgegrid-python==1.0.10
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
govdelivery==1.4.0
Jinja2==2.11.3
lxml==4.6.5
Markdown==3.2.1
ntplib==0.3.4
openpyxl==3.0.3
psycopg2-binary==2.8.6
python-dateutil==2.7.3
regdown==1.0.3
requests==2.25.1
requests_aws4auth==0.8.0
requests_toolbelt==0.8.0
urllib3==1.26.5
wagtail-autocomplete==0.6.3
wagtail-flags==5.1.0
wagtail-inventory==1.4
wagtail-placeholder-images==0.1.1
wagtail-sharing==2.5.1
wagtail-treemodeladmin==1.5.0
wagtailmedia==0.9.0

# These packages are installed from GitHub.
https://github.com/cfpb/owning-a-home-api/releases/download/0.17.2/owning_a_home_api-0.17.2-py3-none-any.whl
https://github.com/cfpb/ccdb5-api/releases/download/1.6.6/ccdb5_api-1.6.6-py3-none-any.whl
https://github.com/cfpb/curriculum-review-tool/releases/download/2.1.7/crtool-2.1.7-py3-none-any.whl

git+https://github.com/cfpb/wagtail-portablepages.git@experiment#egg=wagtail-portablepages

# Temporarily pin Jinja 2.11's MarkupSafe dependency.
MarkupSafe==2.0.1
